requirements.txt
fastapi==0.104.1
uvicorn==0.24.2
streamlit==1.28.1
redis==5.0.1
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
pytest-cov==4.1.0
pyyaml==6.0.1
asyncio-mqtt==0.12.1 # Fallback if needed, but using redis
faker==19.11.0 # For demo data
docker==7.0.0 # For container management